+++ pkg/sources/gitlab/gitlab.go
@@ -64,6 +64,9 @@ type Source struct {
 	sources.CommonSourceUnitUnmarshaller
 
 	useAuthInUrl bool
+
+	clonePath string
+	noCleanup bool
 }
 
 // WithCustomContentWriter sets the useCustomContentWriter flag on the source.
@@ -163,6 +166,8 @@ func (s *Source) Init(ctx context.Context, name string, jobId sources.JobID, sou
 	s.ignoreRepos = conn.GetIgnoreRepos()
 	s.includeRepos = conn.GetIncludeRepos()
 	s.enumerateSharedProjects = !conn.ExcludeProjectsSharedIntoGroups
+	s.clonePath = conn.GetClonePath()
+	s.noCleanup = conn.GetNoCleanup()
 
 	// configuration uses the inverse logic of the `useAuthInUrl` flag.
 	s.useAuthInUrl = !conn.RemoveAuthInUrl
@@ -879,7 +884,7 @@ func (s *Source) scanRepos(ctx context.Context, chunksChan chan *sources.Chunk) 
 			var repo *gogit.Repository
 			var err error
 			if s.authMethod == "UNAUTHENTICATED" {
-				path, repo, err = git.CloneRepoUsingUnauthenticated(ctx, repoURL)
+				path, repo, err = git.CloneRepoUsingUnauthenticated(ctx, repoURL, s.clonePath)
 			} else {
 				// If a username is not provided we need to use a default one in order to clone a private repo.
 				// Not setting "placeholder" as s.user on purpose in case any downstream services rely on a "" value for s.user.
@@ -888,13 +893,17 @@ func (s *Source) scanRepos(ctx context.Context, chunksChan chan *sources.Chunk) 
 					user = "placeholder"
 				}
 
-				path, repo, err = git.CloneRepoUsingToken(ctx, s.token, repoURL, user, s.useAuthInUrl)
+				path, repo, err = git.CloneRepoUsingToken(ctx, s.token, repoURL, s.clonePath, user, s.useAuthInUrl)
 			}
 			if err != nil {
 				scanErrs.Add(err)
 				return nil
 			}
-			defer os.RemoveAll(path)
+
+			// cleanup the path if it is a clone path and --no-cleanup is not set.
+			if !s.noCleanup && s.clonePath != "" {
+				defer os.RemoveAll(path)
+			}
 
 			logger.V(2).Info("starting scan", "num", i+1, "total", len(s.repos))
 			if err = s.git.ScanRepo(ctx, repo, path, s.scanOptions, sources.ChanReporter{Ch: chunksChan}); err != nil {
@@ -1068,7 +1077,7 @@ func (s *Source) ChunkUnit(ctx context.Context, unit sources.SourceUnit, reporte
 	var repo *gogit.Repository
 	var err error
 	if s.authMethod == "UNAUTHENTICATED" {
-		path, repo, err = git.CloneRepoUsingUnauthenticated(ctx, repoURL)
+		path, repo, err = git.CloneRepoUsingUnauthenticated(ctx, repoURL, s.clonePath)
 	} else {
 		// If a username is not provided we need to use a default one in order to clone a private repo.
 		// Not setting "placeholder" as s.user on purpose in case any downstream services rely on a "" value for s.user.
@@ -1077,12 +1086,16 @@ func (s *Source) ChunkUnit(ctx context.Context, unit sources.SourceUnit, reporte
 			user = "placeholder"
 		}
 
-		path, repo, err = git.CloneRepoUsingToken(ctx, s.token, repoURL, user, s.useAuthInUrl)
+		path, repo, err = git.CloneRepoUsingToken(ctx, s.token, repoURL, s.clonePath, user, s.useAuthInUrl)
 	}
 	if err != nil {
 		return err
 	}
-	defer os.RemoveAll(path)
+
+	// cleanup the path if it is a clone path and --no-cleanup is not set.
+	if !s.noCleanup && s.clonePath != "" {
+		defer os.RemoveAll(path)
+	}
 
 	return s.git.ScanRepo(ctx, repo, path, s.scanOptions, reporter)
 }
